Move tutorials section to components
