+++ frontend/src/pages/Tutorials/components/TutorialSection/tutorialsSectionStyle.ts
@@ -1,0 +1,9 @@
+import { styled } from 'goober'
+
+export const CardsList = styled('div')`
+  display: flex;
+  flex-direction: row;
+  gap: 0.5em;
+  flex-wrap: wrap;
+  align-items: flex-start;
+`
