+++ frontend/src/pages/Tutorials/components/index.ts
@@ -1,0 +1,3 @@
+export { default as TutorialCard } from './TutorialCard/TutorialCard'
+export { default as TutorialsSection } from './TutorialSection/TutorialsSection'
+export { default as EmbeddedVideo } from './YoutubeTutorialEmbed/YoutubeTutorialEmbed'
