+++ frontend/src/pages/Tutorials/tutorialsStyle.ts
@@ -10,11 +10,3 @@ export const TitleRow = styled('div')`
 export const Title = styled('h2')`
   margin-block: 0;
 `
-
-export const CardsList = styled('div')`
-  display: flex;
-  flex-direction: row;
-  gap: 0.5em;
-  flex-wrap: wrap;
-  align-items: flex-start;
-`
